Added security to routes

+++ server/api/photos.js
@@ -17,31 +17,22 @@ const upload = multer({ storage }).single('userPhoto');
 
 router.get('/', (req, res, next) => {
   photo
-    .findAll()
-    .then(photos => res.send(photos))
-    .catch(err => next(err));
-});
-
-router.get('/:id', (req, res, next) => {
-  let { id } = req.params;
-  photo
     .findAll({
-      where: { userId: id },
+      where: { userId: req.user.id },
       limit: 6
     })
     .then(photo => res.send(photo))
     .catch(err => next(err));
 });
 
-router.post('/:id', (req, res, next) => {
-  let user = req.params.id;
+router.post('/', (req, res, next) => {
   upload(req, res, err => {
     if (!req.file) {
       next(err);
     } else {
       photo
         .create({ path: req.file.path, destination: req.file.path })
-        .then(newPhoto => newPhoto.setUser(user))
+        .then(newPhoto => newPhoto.setUser(req.user.id))
         .then(x => res.send(x))
         .catch(err => next(err));
     }
@@ -50,20 +41,22 @@ router.post('/:id', (req, res, next) => {
 
 router.delete('/:id', (req, res, next) => {
   let { id } = req.params;
-  photo
-    .findOne({
-      where: { id }
-    })
-    .then(deleteFromFs => {
-      fs.unlink(deleteFromFs.destination, err => {
-        if (err) throw err;
-        console.log(deleteFromFs.destination + ' was DELETED!!!!');
-      });
-      return deleteFromFs;
-    })
-    .then(photo => photo.destroy())
-    .then(() => res.sendStatus(200))
-    .catch(err => next(err));
+  if (req.user) {
+    photo
+      .findOne({
+        where: { id }
+      })
+      .then(deleteFromFs => {
+        fs.unlink(deleteFromFs.destination, err => {
+          if (err) throw err;
+          console.log(deleteFromFs.destination + ' was DELETED!!!!');
+        });
+        return deleteFromFs;
+      })
+      .then(photo => photo.destroy())
+      .then(() => res.sendStatus(200))
+      .catch(err => next(err));
+  }
 });
 
 module.exports = router;
